Moving feature tests to a single file.
